requirements.txt
aiogram>=3.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.2.2
torch>=2.0.1
PyPDF2>=3.0.1
python-docx>=0.8.11
stop-words
nltk~=3.9.1